fully return result struct
