+++ pkg/controller/issueapi/show_bulk_issue.go
